Authorization/Error handling

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require('express');
 const multer = require('multer');
 
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
 
@@ -63,12 +64,13 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.post('', multer({ storage: storage }).single('image'), (req, res, next) => {
+router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
   const url = req.protocol + '://' + req.get('host');
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId,
   });
   post.save().then(result=>{
     res.status(201).json({
@@ -81,14 +83,18 @@ router.post('', multer({ storage: storage }).single('image'), (req, res, next) =
   });
 });
 
-router.delete('/:id', (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+router.delete('/:id', checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: 'Post deleted!' });
+    if (result.n) {
+      res.status(200).json({ message: 'Post deleted!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
   });
 });
 
-router.put('/:id', multer({ storage: storage }).single('image'), (req, res, next) => {
+router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
   let imagePath = req.body.imagePath;
   if (req.file) {
     const url = req.protocol + '://' + req.get('host');
@@ -100,9 +106,12 @@ router.put('/:id', multer({ storage: storage }).single('image'), (req, res, next
     content: req.body.content,
     imagePath: imagePath,
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Post updated!' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified) {
+      res.status(200).json({ message: 'Post updated!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
   })
 });
